modularizacion y doc
