+++ include/pins.h
@@ -13,6 +13,8 @@
 /**
  * PORT definitions
  */
+#pragma once
+
 #define LED_PORT GPIOA
 #define BUZZER_PORT GPIOA
 #define BUTTON_PORT GPIOA
@@ -21,20 +23,19 @@
 /**
  * PIN definitions 
  */
+
 #define YELLOW_LED_PIN GPIO12
 #define RED_LED_PIN GPIO9
 #define GREEN_LED_PIN GPIO8
 #define BUZZER_PIN GPIO3       
 #define BUTTON_PIN GPIO0
+#define LED_TX  GPIO10          
 #define ADC_PIN_vib GPIO1
 #define ADC_PIN_hum GPIO2
-#define LED_TX  GPIO10          //led indicador de envio de datos
-/**
- * ADC channel definitions
- */
-#define ADC_CHANNEL_hum ADC_CHANNEL1
-#define ADC_CHANNEL_vib ADC_CHANNEL2
-
-#define ADC_BUFFER_SIZE 32
 
 
+/*
+Functions prototypes ------------------------------------------------------
+*/
+
+void gpio_setup(void);
