--- src/util.c
@@ -1,6 +1,0 @@
-#include "../include/util.h"
-#include "../include/pins.h"
-#include <libopencm3/stm32/rcc.h>
-#include <libopencm3/stm32/gpio.h>
-#include <libopencm3/stm32/adc.h>
-
